adding changes suggested by mentor
