+++ src/components/TodoList.tsx
@@ -1,28 +1,128 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
-import React from 'react';
+import React, { useState } from 'react';
 import { Todo } from '../types/Todo';
 import { CSSTransition, TransitionGroup } from 'react-transition-group';
+import { deleteTodo, updateTodo } from '../api/todos';
 
 interface TodoListProps {
   filteredTodos: Todo[];
-  handleUpdateTodoStatus: (id: number) => void;
-  handleRemoveTodo: (id: number) => void;
-  handleSelectedTodoChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  handleSubmitForm: (e: React.FormEvent<HTMLFormElement>, todo: Todo) => void;
-  editingTodoId: number[];
-  selectedTodo: Todo | null;
-  setSelectedTodo: (todo: Todo | null) => void;
+  editingTodosId: number[];
+  setEditingTodosId: (ids: number[]) => void;
+  setError: (error: string) => void;
+  todos: Todo[];
+  setTodos: (todos: Todo[]) => void;
+  setFilteredTodos: (todos: Todo[]) => void;
+  filter: string;
 }
 export function TodoList({
+  todos,
   filteredTodos,
-  handleUpdateTodoStatus,
-  handleRemoveTodo,
-  handleSelectedTodoChange,
-  handleSubmitForm,
-  editingTodoId,
-  selectedTodo,
-  setSelectedTodo,
+  editingTodosId,
+  setEditingTodosId,
+  setError,
+  setTodos,
+  setFilteredTodos,
 }: TodoListProps) {
+  const [selectedTodo, setSelectedTodo] = useState<Todo | null>(null);
+
+  function handleSelectedTodoChange(e: React.ChangeEvent<HTMLInputElement>) {
+    if (!selectedTodo) {
+      return;
+    }
+
+    setSelectedTodo({
+      ...selectedTodo,
+      title: e.target.value,
+    });
+  }
+
+  async function handleEditTitleOfTodo(todo: Todo) {
+    try {
+      if (!selectedTodo) {
+        throw new Error('Unable to update todo');
+      }
+
+      setEditingTodosId([...editingTodosId, todo.id]);
+      const updatedTodoFromServer = await updateTodo({
+        ...todo,
+        title: selectedTodo.title,
+      });
+      const newTodos = todos.map(t =>
+        t.id === todo.id ? updatedTodoFromServer : t,
+      );
+
+      setTodos(newTodos);
+      setFilteredTodos(newTodos);
+      setEditingTodosId(editingTodosId.filter(id => id !== todo.id));
+    } catch (e) {
+      setError('Unable to update a todo');
+      setEditingTodosId(editingTodosId.filter(id => id !== todo.id));
+    }
+  }
+
+  async function handleUpdateTodoStatus(todoId: number) {
+    try {
+      const todo = todos.find(t => t.id === todoId);
+
+      setEditingTodosId([...editingTodosId, todoId]);
+      if (!todo) {
+        throw new Error('Todo not found');
+      }
+
+      const updatedTodo: Todo = { ...todo, completed: !todo.completed };
+      const updatedTodoFromServer = await updateTodo(updatedTodo);
+      const newTodos = todos.map(t =>
+        t.id === todoId ? updatedTodoFromServer : t,
+      );
+
+      setEditingTodosId(editingTodosId.filter(id => id !== todoId));
+      setTodos(newTodos);
+      setFilteredTodos(newTodos);
+    } catch (e) {
+      setError('Unable to update a todo');
+      setEditingTodosId(editingTodosId.filter(id => id !== todoId));
+    }
+  }
+
+  async function handleRemoveTodo(todoId: number) {
+    try {
+      setEditingTodosId([...editingTodosId, todoId]);
+      await deleteTodo(todoId);
+      setEditingTodosId(editingTodosId.filter(id => id !== todoId));
+      setTodos(todos.filter(t => t.id !== todoId));
+      setFilteredTodos(filteredTodos.filter(t => t.id !== todoId));
+    } catch (e) {
+      setError('Unable to delete a todo');
+      setEditingTodosId(editingTodosId.filter(id => id !== todoId));
+    }
+  }
+
+  async function handleEditTodoTitle(
+    e: React.FormEvent<HTMLFormElement>,
+    todo: Todo,
+  ) {
+    e.preventDefault();
+    if (!selectedTodo) {
+      return;
+    }
+
+    if (selectedTodo && !selectedTodo.title) {
+      await handleRemoveTodo(todo.id);
+      setSelectedTodo(null);
+
+      return;
+    } else if (selectedTodo.title === todo.title) {
+      setSelectedTodo(null);
+
+      return;
+    } else if (selectedTodo.title !== todo.title) {
+      await handleEditTitleOfTodo(selectedTodo);
+      setSelectedTodo(null);
+
+      return;
+    }
+  }
+
   return (
     <TransitionGroup component={null}>
       {filteredTodos.map(todo => (
@@ -58,7 +158,7 @@ export function TodoList({
                 {todo.title}
               </span>
             ) : (
-              <form onSubmit={e => handleSubmitForm(e, todo)}>
+              <form onSubmit={e => handleEditTodoTitle(e, todo)}>
                 <input
                   data-cy="TodoTitleField"
                   type="text"
@@ -66,7 +166,7 @@ export function TodoList({
                   onChange={handleSelectedTodoChange}
                   placeholder="Empty todo will be deleted"
                   onBlur={e =>
-                    handleSubmitForm(
+                    handleEditTodoTitle(
                       e as unknown as React.FormEvent<HTMLFormElement>,
                       todo,
                     )
@@ -89,7 +189,7 @@ export function TodoList({
             {/* overlay will cover the todo while it is being deleted or updated */}
             <div
               data-cy="TodoLoader"
-              className={`modal overlay ${editingTodoId.includes(todo.id) ? 'is-active' : ''}`}
+              className={`modal overlay ${editingTodosId.includes(todo.id) ? 'is-active' : ''}`}
             >
               <div className="modal-background has-background-white-ter" />
               <div className="loader" />
